+++ lib/runit.rb
@@ -18,7 +18,7 @@ module Runit
   SERVICES_DIR = Pathname.new(__dir__).join("../services").expand_path
   LOG_DIR = Pathname.new(__dir__).join("../log").expand_path
 
-  ALL_DATA_ORIENTED_SERVICE_NAMES = %w[minio openldap gitaly praefect redis postgresql-geo postgresql].freeze
+  ALL_DATA_ORIENTED_SERVICE_NAMES = %w[minio openldap gitaly praefect redis redis-cluster postgresql-geo postgresql].freeze
   STOP_RETRY_COUNT = 3
 
   def self.start_runsvdir